The counter is an int instead of a long
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@84603 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acsGUIs/alarmPanel/src/alma/acsplugins/alarmsystem/gui/AlarmCounter.java
@@ -26,13 +26,13 @@ package alma.acsplugins.alarmsystem.gui;
  */
 public class AlarmCounter {
 	
-	private volatile long count=0;
+	private volatile int count=0;
 	
 	/**
 	 * Increase the counter
 	 */
 	public void incCounter() {
-		if (count==Long.MAX_VALUE) {
+		if (count==Integer.MAX_VALUE) {
 			count=1;
 			return;
 		}
